+++ src/api/api.ts
@@ -8,7 +8,7 @@ import axios, {
     CancelToken
 } from "axios";
 import Permission from "../model/permission";
-import { IFacets } from "../components/browse-files/FileFilter";
+import { IFacets } from "../components/browse-data/shared/useFilterFacets";
 
 const URL: string = process.env.REACT_APP_API_URL!;
 
@@ -118,10 +118,10 @@ function getAccountInfo(token: string): Promise<Account | undefined> {
         .then(_extractItem);
 }
 
-function getTrials(token: string): Promise<Trial[]> {
+function getTrials(token: string, params: any = {}): Promise<Trial[]> {
     return getApiClient(token)
         .get("trial_metadata", {
-            params: { sort_field: "trial_id", sort_direction: "asc" }
+            params: { sort_field: "trial_id", sort_direction: "asc", ...params }
         })
         .then(_extractItems);
 }
