+++ src/components/browse-data/files/FileDetails.test.tsx
