+++ src/components/browse-data/trials/TrialTable.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const TrialTable: React.FC = () => {
+    return <>trial table</>;
+};
+
+export default TrialTable;
